Initial POM cleanup
This cleans up the POM files in several ways:

Dependency versions are declared in the parent POM so that all version
updates happen in one place.

The user is now required to use JDK 9 to compile the code, but the
produced bytecode is guaranteed to be compatible with JDK 7 and above
thanks to the use of the -release flag.

Plugin versions are replaced with properties for painless updates with
the Maven Versions plugin.

All POM files were run through Intellij IDEA's XML formatter for
consistent indentation.

Compiler severity has been reduced from -Werror temporarily due to
new warnings that would have otherwise broken the build.