Attempt to modularize the project
This is another attempt to modularize the project. This version
should provide better support for building the code on older
compilers (Android in particular).

It works by adding a profile named "on-jdk-9-plus" that is activated
when the building JDK is >= JDK 9. This profile adds Java 9-specific
source directories named src/main/java9 to each module containing
a module descriptor. The profile also upgrades SLF4J to a Java 9
compatible version.

It also adds other JDKs to the Travis CI configuration to test that
the project builds on all of them.