Enable OSGi support
This adds the relevant plugin configurations to make ode4j compatible
with the OSGi module system [0]. These are solely updates to the Jar
file manifests and therefore pose no compatibility issues. Manifest
generation is achieved via the bnd-maven-plugin [1] due to this plugin
having slightly better support across all of the major IDEs than other
plugins such as the maven-bundle-plugin [2].

In the process, it removes a couple of spurious compiler executions
from submodules.

[0] https://en.wikipedia.org/wiki/OSGi
[1] https://github.com/bndtools/bnd/tree/master/maven/bnd-maven-plugin
[2] https://felix.apache.org/documentation/subprojects/apache-felix-maven-bundle-plugin-bnd.html